fastapi_backend/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
transformers==4.47.0  # https://huggingface.co/transformers/
scikit-learn==1.5.2  # https://scikit-learn.org/stable/
torch==2.5.0  # https://pytorch.org/
transformers[torch]==4.47.0   # https://huggingface.co/transformers/
datasets==3.2.0  # https://huggingface.co/docs/datasets/

tqdm~=4.67.1
pandas~=2.2.3
Faker~=33.1.0
babel~=2.16.0
numpy~=2.1.3

psycopg2-binary